requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dateutil==2.9.0
